Update files to referance Coverage capability
Update SetupTestingEnvironment.rst and ReviewingResults.rst to instruct
the user on how to enable and find the coverage output on CDash.

Add figures for the USE_COVERAGE option and the warnings show to
SetupTestingEvironment.  Add figures for showing the Coverage group on the
dashboard and an example output for one routine.