Revise index pages
Add the file sphinx.rst to serve as the index page for the Sphinx
generated documentation.  This should contain the same code as the old
index.rst

Rewrite the index.rst to be the index for the Github rendered
documentation.

Change the conf.py to reference the sphinx.rst as the toc-tree master_doc