Fix broken image paths and fix text.
Separate the paragraph about the system management portal and the
following figure with a blank line so that the figure markup can be
utilized.

Fix the link for the Cmake2011GlobalMappingSelect to have the full eight
character hash snippet.  A '0'(zero) needs to be added to the end of the
hash for the path to be correct.

Remove extra ':alt:' reference found in the sentence about the Cache
licensing window.